requirements.txt
python-telegram-bot
psycopg2-binary
python-dotenv
apscheduler
asyncio
ethiopian-date-converter
#ethiopian_date
#telegram